requirements.txt
discord.py==2.0.1
async-timeout==4.0.2
python-dotenv==0.21.0
python-dateutil==2.8.2
jsonlib-python3==1.6.1
requests==2.28.1
urllib3==1.26.9

